Fix lint-related errors
