+++ datajunction-server/.pylintrc
@@ -1,4 +1,5 @@
 [MESSAGES CONTROL]
+disable=cyclic-import
 
 [MASTER]
 # https://github.com/samuelcolvin/pydantic/issues/1961#issuecomment-759522422
