+++ datajunction-server/datajunction_server/api/access/authentication/google.py
@@ -1,3 +1,4 @@
+# pylint: disable=cyclic-import
 """
 Google OAuth Router
 """
